fix verify_socket() function and remove check for errno
we can only use verify_socket() when we actually set the socket to non
blocking so it's now moved there.
there where also mutiple changes in it that caused it to break in
horrible ways.

also removed checks for errno as this should not be checked for.
errno will only be set when an error accured so checking for it without
first having a succesfull check for -1 is undefined behaviour.
errno maybe be set in some internal functions and been dealt with and
stay over.