Implement IRC hooks and move most code to it
This is done through a global GHashTable to hook commands to a linked
list of functions to call. A special PREINIT command used to allow for
executing functions before registering the client. However, since client
registration is done in PREINIT, it has to be the last element in the
linked list. This is only done by hooking it after every other PREINIT
hook. exec_hooks runs PREINIT hooks when msg is NULL, it runs "*" hooks
when msg is 1, else it grabs the command from the message and runs all
hooks for that command. "*" hooks run for every single message. This is
to allow something like an module engine to run its own hooks interface.

This allowed a lot of code to be moved from main and the callback to a
file containing core hooks. It also makes irc_init_loop_callback
unnecessary.

We can use this to start developing modules in C, although the plan is
to use a scheme interpreter and to provide it with an API for developing
modules.

* Implement fully async I/O through libev, and some fixes
* Message processing is now queue-based
* Add a timeout watcher to the event loop
* Push message processing to handle_message, outside of the event loop
* Changed check_socket for verify_socket
* Fix some warnings
* Push the call to setnonblock to after the TLS setup to avoid an error
* Check whether the socket is invalid only after we're done looping over addrinfo
* Check for errno after irc_server_connect
* Add basic a basic IRC hooks API
* Finalize the hooks API and add hooks to it, simplifying init code
* Get rid of irc_init_loop_callback, use hooks instead
* Add core_hooks.c
* Make a hook that matches every IRC message
* Run clang-format on everything
* Uncomment the channel autojoin hook
* Add sanity checks in handle_message
* Make the IrciumMessage constant when sending to hooks for sanity
* set gbuf to NULL
* Use g_str_equal instead of strcmp