switch out ircium parser for new ircmsg parser
removes the old ircium parser completly and replaces it with the new
more flexible ircmsg parser, that also get's rid of most of the glib.

New irc_msg_new function to create a new irc_msg object. takes optionaly
the prefix, a command, the number of parameters, and an array of
parameters.